fix: corrects selectRoomDisplayStates based on copilot recommendation and moves to selectors file
