+++ src/lib/store/rooms/rooms.selectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { RootState } from 'src/lib';
+import { DisplayState, RootState, selectAllDevices } from 'src/lib';
 
 
 const roomsState = (state: RootState) => state.rooms;
@@ -106,4 +106,30 @@ export const selectRoomShareState = (roomKey: string) => createSelector(
 export const selectZoomRoomControllerKey = (roomKey: string) => createSelector(
   roomsState,
   (rooms) => rooms[roomKey] ? rooms[roomKey]?.configuration?.zoomRoomControllerKey : undefined
+);
+
+/**
+ * Get the display states for the room
+ * Exludes the programAudio and codecContent destinations
+ * @param roomKey
+ * @returns the display states for the room's displays
+ */
+export const selectRoomDisplayStates = createSelector(
+  [
+    (state: RootState, roomKey: string) => roomKey,
+    () => selectAllDevices,
+    (state: RootState, roomKey: string) => state.rooms[roomKey]?.configuration?.destinations,
+  ],
+  (roomKey, deviceStates, destinations) => {
+    console.log("roomKey", roomKey);
+    console.log("devices", deviceStates);
+    console.log("destinations", destinations);
+    if (!destinations) return undefined;
+
+    const displayKeys = Object.entries(destinations).filter(([key]) => key !== "programAudio" && key !== "codecContent").map(([, value]) => value);
+
+    const displayStates = Object.values(deviceStates).filter((device) => Object.values(displayKeys).includes(device.key));
+
+    return displayStates as DisplayState[];
+  }
 );
